CRUD tests

+++ includes/funcionesTests.php
@@ -8,34 +8,41 @@ function obtenerPreguntasTestParaAdmin($preguntas){
         $html.="</tr>";
     }
     $html.="</table>";
+    return $html;
 }
 
 function obtenerPreguntaParaAdmin($pregunta){
-    $html="<td>";
-    $html.=$pregunta->getIDPregunta();
-    $html.="</td>";
-    $html.="<td>";
-    $html.=$pregunta->getPregunta();
-    $html.="</td>";
-    $html.="<td>";
-    $html.=$pregunta->getRespuestaCorrecta();
-    $html.="</td>";
-    $html.="<td>";
-    $html.=$pregunta->getRespuesta1();
-    $html.="</td>";
-    $html.="<td>";
-    $html.=$pregunta->getRespuesta2();
-    $html.="</td>";
-    $html.="<td>";
-    $html.=$pregunta->getRespuesta3();
-    $html.="</td>";
-    $html.="<td>";
-    $html.="<a href='editarPregunta.php?id=";
-    $html.=$pregunta->getIDPregunta();
-    $html.="'><button>Editar</button></a>";
-    $html.="<a href='eliminarPregunta.php?id=";
-    $html.=$pregunta->getIDPregunta();
-    $html.="'><button>Eliminar</button></a>";
-    $html.="</td>";
+    if ($pregunta->getIDPregunta()!=null){
+        $html="<td>";
+        $html.=$pregunta->getIDPregunta();
+        $html.="</td>";
+        $html.="<td>";
+        $html.=$pregunta->getPregunta();
+        $html.="</td>";
+        $html.="<td>";
+        $html.=$pregunta->getRespuestaCorrecta();
+        $html.="</td>";
+        $html.="<td>";
+        $html.=$pregunta->getRespuesta1();
+        $html.="</td>";
+        $html.="<td>";
+        $html.=$pregunta->getRespuesta2();
+        $html.="</td>";
+        $html.="<td>";
+        $html.=$pregunta->getRespuesta3();
+        $html.="</td>";
+        $html.="<td>";
+        $html.="<a href='editarPregunta.php?id=";
+        $html.=$pregunta->getIDPregunta();
+        $html.="&idTest=";
+        $html.=$_GET["id"];
+        $html.="'><button>Editar</button></a>";
+        $html.="<a href='eliminarPregunta.php?id=";
+        $html.=$pregunta->getIDPregunta();
+        $html.="&idTest=";
+        $html.=$_GET["id"];
+        $html.="'><button>Eliminar</button></a>";
+        $html.="</td>";
+    }
     return $html;
 }
